Move demo icons to demo folders.
