+++ packages/ui/src/components/vuestic-components/va-icon/demo/VaIconVue.vue
