+++ packages/ui/src/components/vuestic-components/va-navbar/demo/VaIconMenuCollapsed.vue
